feat(tts): refactor TTS architecture with new helper classes and services
