--- android/src/main/java/com/sherpaonnx/SherpaOnnxTtsHelper.kt
@@ -1,1412 +1,0 @@
-package com.sherpaonnx
-
-import android.app.ActivityManager
-import android.content.Context
-import android.media.AudioAttributes
-import android.media.AudioFormat
-import android.media.AudioManager
-import android.media.AudioTrack
-import android.net.Uri
-import android.os.Build
-import android.os.Handler
-import android.os.Looper
-import android.util.Log
-import com.facebook.react.bridge.Arguments
-import com.facebook.react.bridge.Promise
-import com.facebook.react.bridge.ReadableArray
-import com.facebook.react.bridge.ReadableMap
-import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.bridge.WritableMap
-import com.k2fsa.sherpa.onnx.GeneratedAudio
-import com.k2fsa.sherpa.onnx.GenerationConfig
-import com.k2fsa.sherpa.onnx.OfflineTts
-import com.k2fsa.sherpa.onnx.OfflineTtsConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsPocketModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsVitsModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsMatchaModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsKokoroModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsKittenModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsZipVoiceModelConfig
-import com.k2fsa.sherpa.onnx.OfflineTtsSupertonicModelConfig
-import java.io.File
-import java.io.FileInputStream
-import java.io.FileOutputStream
-import java.io.OutputStream
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.Executors
-import java.util.concurrent.atomic.AtomicBoolean
-
-internal class SherpaOnnxTtsHelper(
-  private val context: ReactApplicationContext,
-  private val detectTtsModel: (modelDir: String, modelType: String) -> HashMap<String, Any>?,
-  private val emitChunk: (String, String, FloatArray, Int, Float, Boolean) -> Unit,
-  private val emitError: (String, String, String) -> Unit,
-  private val emitEnd: (String, String, Boolean) -> Unit,
-  /** FFmpeg: mono f32le raw file → encoded output path. Returns empty string on success. */
-  private val encodeMonoFromRawFile: (rawPath: String, pcmSampleRate: Int, outputPath: String, format: String, outputSampleRateHz: Int) -> String
-) {
-
-  private data class TtsInitState(
-    val modelDir: String,
-    val modelType: String,
-    val numThreads: Int,
-    val debug: Boolean,
-    val noiseScale: Double?,
-    val noiseScaleW: Double?,
-    val lengthScale: Double?,
-    val ruleFsts: String?,
-    val ruleFars: String?,
-    val maxNumSentences: Int?,
-    val silenceScale: Double?,
-    val provider: String?
-  )
-
-  private data class TtsEngineInstance(
-    @Volatile var tts: OfflineTts? = null,
-    @Volatile var ttsInitState: TtsInitState? = null,
-    val ttsStreamRunning: AtomicBoolean = AtomicBoolean(false),
-    val ttsStreamCancelled: AtomicBoolean = AtomicBoolean(false),
-    var ttsStreamThread: Thread? = null,
-    var ttsPcmTrack: AudioTrack? = null
-  ) {
-    private val lock = Any()
-
-    fun hasEngine(): Boolean = synchronized(lock) { tts != null }
-    val isZipvoice: Boolean get() = ttsInitState?.modelType == "zipvoice"
-    val isPocket: Boolean get() = ttsInitState?.modelType == "pocket"
-    fun releaseEngines() {
-      synchronized(lock) {
-        tts?.release()
-        tts = null
-        ttsInitState = null
-      }
-    }
-    fun stopPcmPlayer() {
-      synchronized(lock) {
-        ttsPcmTrack?.apply {
-          try { stop() } catch (_: IllegalStateException) {}
-          flush()
-          release()
-        }
-        ttsPcmTrack = null
-      }
-    }
-  }
-
-  private val instances = ConcurrentHashMap<String, TtsEngineInstance>()
-
-  private fun getInstance(instanceId: String): TtsEngineInstance? = instances[instanceId]
-
-  /** Run promise resolve/reject on the UI thread so React state updates run on the main thread. */
-  private val mainHandler = Handler(Looper.getMainLooper())
-  private fun resolveOnUiThread(promise: Promise, result: WritableMap) {
-    mainHandler.post { promise.resolve(result) }
-  }
-  private fun rejectOnUiThread(promise: Promise, code: String, message: String, throwable: Throwable? = null) {
-    mainHandler.post {
-      if (throwable != null) promise.reject(code, message, throwable) else promise.reject(code, message)
-    }
-  }
-
-  /**
-   * libsherpa-onnx-jni looks up `invoke([F)Ljava/lang/Integer` (see sherpa-onnx `offline-tts.cc` CallCallback).
-   * Kotlin `Function1<*, Int>` compiles to `invoke([F)I`, so GetMethodID fails and JNI aborts.
-   * Using [java.lang.Integer] as the type parameter yields the boxed JVM signature the JNI expects.
-   * The cast is only for the Kotlin API (`generateWithCallback` still declares `Function1<FloatArray, Int>`).
-   */
-  /** Box for JNI: must be real [java.lang.Integer], not Kotlin [Int] (primitive `invoke([F)I` breaks sherpa JNI). */
-  @Suppress("DEPRECATION")
-  private fun boxForTtsJni(n: Int): java.lang.Integer = java.lang.Integer(n)
-
-  @Suppress("UNCHECKED_CAST")
-  private fun ttsChunkCallbackForJni(
-    sentenceChunkSizes: MutableList<Int>
-  ): kotlin.Function1<FloatArray, Int> {
-    val boxed =
-      object : kotlin.jvm.functions.Function1<FloatArray, java.lang.Integer> {
-        override fun invoke(chunk: FloatArray): java.lang.Integer {
-          sentenceChunkSizes.add(chunk.size)
-          return boxForTtsJni(chunk.size)
-        }
-      }
-    return boxed as kotlin.Function1<FloatArray, Int>
-  }
-
-  @Suppress("UNCHECKED_CAST")
-  private fun ttsStreamChunkCallbackForJni(
-    cancelled: AtomicBoolean,
-    onChunk: (FloatArray) -> Unit
-  ): kotlin.Function1<FloatArray, Int> {
-    val boxed =
-      object : kotlin.jvm.functions.Function1<FloatArray, java.lang.Integer> {
-        override fun invoke(chunk: FloatArray): java.lang.Integer {
-          if (cancelled.get()) return boxForTtsJni(0)
-          onChunk(chunk)
-          return boxForTtsJni(chunk.size)
-        }
-      }
-    return boxed as kotlin.Function1<FloatArray, Int>
-  }
-
-  /** Single-thread executor for TTS init so the RN bridge thread is not blocked (avoids Inspector/dev WebSocket races in debug builds). */
-  private val ttsInitExecutor = Executors.newSingleThreadExecutor()
-
-  /**
-   * Shuts down the TTS init executor and releases all engine instances.
-   * Call from the native module's onCatalystInstanceDestroy() to avoid leaking the executor thread.
-   */
-  fun shutdown() {
-    try {
-      ttsInitExecutor.shutdown()
-      if (!ttsInitExecutor.awaitTermination(3, java.util.concurrent.TimeUnit.SECONDS)) {
-        ttsInitExecutor.shutdownNow()
-      }
-    } catch (e: InterruptedException) {
-      Thread.currentThread().interrupt()
-      ttsInitExecutor.shutdownNow()
-    }
-    instances.values.forEach { inst ->
-      inst.releaseEngines()
-      inst.stopPcmPlayer()
-    }
-    instances.clear()
-  }
-
-  fun initializeTts(
-    instanceId: String,
-    modelDir: String,
-    modelType: String,
-    numThreads: Double,
-    debug: Boolean,
-    noiseScale: Double?,
-    noiseScaleW: Double?,
-    lengthScale: Double?,
-    ruleFsts: String?,
-    ruleFars: String?,
-    maxNumSentences: Double?,
-    silenceScale: Double?,
-    provider: String?,
-    promise: Promise
-  ) {
-    ttsInitExecutor.execute init@{
-      try {
-      val result = detectTtsModel(modelDir, modelType)
-      if (result == null) {
-        Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: Failed to detect TTS model: native call returned null")
-        rejectOnUiThread(promise, "TTS_INIT_ERROR", "Failed to detect TTS model: native call returned null")
-        return@init
-      }
-      val success = result["success"] as? Boolean ?: false
-      if (!success) {
-        val reason = result["error"] as? String
-        Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: ${reason ?: "Failed to detect TTS model"}")
-        rejectOnUiThread(promise, "TTS_INIT_ERROR", reason ?: "Failed to detect TTS model")
-        return@init
-      }
-      val paths = (result["paths"] as? Map<*, *>)?.mapValues { (_, v) -> (v as? String).orEmpty() }?.mapKeys { it.key.toString() } ?: emptyMap()
-      val modelTypeStr = result["modelType"] as? String ?: "vits"
-      val detectedModels = result["detectedModels"] as? ArrayList<*>
-
-      val inst = instances.getOrPut(instanceId) { TtsEngineInstance() }
-      inst.stopPcmPlayer()
-      inst.releaseEngines()
-
-      val sampleRate: Int
-      val numSpeakers: Int
-
-      if (modelTypeStr == "zipvoice") {
-        val vocoderPath = path(paths, "vocoder")
-        if (vocoderPath.isBlank()) {
-          val msg = "Zipvoice distill models (encoder+decoder only, no vocoder) are not supported. Use the full Zipvoice model that includes vocos_24khz.onnx (or similar vocoder file)."
-          Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: $msg")
-          rejectOnUiThread(promise, "TTS_INIT_ERROR", msg)
-          return@init
-        }
-        val lexiconPath = path(paths, "lexicon")
-        if (lexiconPath.isBlank()) {
-          val msg = "Zipvoice requires lexicon.txt (or lexicon-<lang>.txt) in the model directory. The sherpa-onnx engine aborts if it is missing. Copy lexicon from the official k2-fsa sherpa-onnx Zipvoice model package or hr-files release next to tokens.txt."
-          Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: $msg")
-          rejectOnUiThread(promise, "TTS_INIT_ERROR", msg)
-          return@init
-        }
-        val am = context.applicationContext.getSystemService(Context.ACTIVITY_SERVICE) as? ActivityManager
-        if (am != null) {
-          val memInfo = ActivityManager.MemoryInfo()
-          am.getMemoryInfo(memInfo)
-          val availMb = memInfo.availMem / (1024 * 1024)
-          if (memInfo.availMem < 800L * 1024 * 1024) {
-            val msg = "Not enough free memory to load the Zipvoice model (available: ${availMb} MB). Close other apps to free memory or use a smaller Zipvoice model that includes all required components (encoder, decoder, and vocoder)."
-            Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: $msg")
-            rejectOnUiThread(promise, "TTS_INIT_ERROR", msg)
-            return@init
-          }
-        }
-        // Hint GC before heavy allocation to reduce memory pressure; zipvoice always uses 1 thread to limit peak RAM.
-        System.gc()
-        if (am != null) {
-          val memInfoBefore = ActivityManager.MemoryInfo()
-          am.getMemoryInfo(memInfoBefore)
-          Log.i("SherpaOnnxTts", "Zipvoice init: availMem=${memInfoBefore.availMem / (1024 * 1024)} MB (before load)")
-        }
-        val zipvoiceNumThreads = 1
-        val config = buildTtsConfig(
-          paths, "zipvoice", zipvoiceNumThreads, debug,
-          noiseScale, noiseScaleW, lengthScale,
-          ruleFsts, ruleFars, maxNumSentences?.toInt(), silenceScale,
-          provider
-        )
-        if (am != null) {
-          val memInfo = ActivityManager.MemoryInfo()
-          am.getMemoryInfo(memInfo)
-          Log.i("SherpaOnnxTts", "Zipvoice init: availMem=${memInfo.availMem / (1024 * 1024)} MB (after load)")
-        }
-        try {
-          inst.tts = OfflineTts(config = config)
-        } catch (e: Exception) {
-          Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: Failed to create Zipvoice OfflineTts: ${e.message}", e)
-          rejectOnUiThread(promise, "TTS_INIT_ERROR", "Failed to create Zipvoice TTS engine: ${e.message}", e)
-          return@init
-        }
-        sampleRate = inst.tts!!.sampleRate()
-        numSpeakers = inst.tts!!.numSpeakers()
-      } else {
-        val config = buildTtsConfig(
-          paths, modelTypeStr, numThreads.toInt(), debug,
-          noiseScale, noiseScaleW, lengthScale,
-          ruleFsts, ruleFars, maxNumSentences?.toInt(), silenceScale,
-          provider
-        )
-        inst.tts = OfflineTts(config = config)
-        sampleRate = inst.tts!!.sampleRate()
-        numSpeakers = inst.tts!!.numSpeakers()
-      }
-
-      val modelsArray = Arguments.createArray()
-      detectedModels?.forEach { modelObj ->
-        if (modelObj is HashMap<*, *>) {
-          val modelMap = Arguments.createMap()
-          modelMap.putString("type", modelObj["type"] as? String ?: "")
-          modelMap.putString("modelDir", modelObj["modelDir"] as? String ?: "")
-          modelsArray.pushMap(modelMap)
-        }
-      }
-
-      inst.ttsInitState = TtsInitState(
-        modelDir,
-        modelTypeStr,  // detected model type (e.g. "pocket"), not the requested "auto"
-        numThreads.toInt(),
-        debug,
-        noiseScale?.takeUnless { it.isNaN() },
-        noiseScaleW?.takeUnless { it.isNaN() },
-        lengthScale?.takeUnless { it.isNaN() },
-        ruleFsts?.takeIf { it.isNotBlank() },
-        ruleFars?.takeIf { it.isNotBlank() },
-        maxNumSentences?.toInt()?.takeIf { it > 0 },
-        silenceScale?.takeUnless { it.isNaN() },
-        provider?.takeIf { it.isNotBlank() }
-      )
-
-      Log.i("SherpaOnnxTts", "initializeTts: instanceId=$instanceId, engine=kotlin-api modelType=$modelTypeStr, sampleRate=$sampleRate, numSpeakers=$numSpeakers")
-
-      val resultMap = Arguments.createMap()
-      resultMap.putBoolean("success", true)
-      resultMap.putArray("detectedModels", modelsArray)
-      resultMap.putInt("sampleRate", sampleRate)
-      resultMap.putInt("numSpeakers", numSpeakers)
-      resolveOnUiThread(promise, resultMap)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_INIT_ERROR: Failed to initialize TTS: ${e.message}", e)
-      rejectOnUiThread(promise, "TTS_INIT_ERROR", "Failed to initialize TTS: ${e.message}", e)
-    }
-    }
-  }
-
-  fun updateTtsParams(
-    instanceId: String,
-    noiseScale: Double?,
-    noiseScaleW: Double?,
-    lengthScale: Double?,
-    promise: Promise
-  ) {
-    val inst = getInstance(instanceId) ?: run {
-      Log.e("SherpaOnnxTts", "TTS_UPDATE_ERROR: TTS instance not found: $instanceId")
-      promise.reject("TTS_UPDATE_ERROR", "TTS instance not found: $instanceId")
-      return
-    }
-    if (inst.ttsStreamRunning.get()) {
-      Log.e("SherpaOnnxTts", "TTS_UPDATE_ERROR: Cannot update params while streaming")
-      promise.reject("TTS_UPDATE_ERROR", "Cannot update params while streaming")
-      return
-    }
-    val state = inst.ttsInitState ?: run {
-      Log.e("SherpaOnnxTts", "TTS_UPDATE_ERROR: TTS not initialized")
-      promise.reject("TTS_UPDATE_ERROR", "TTS not initialized")
-      return
-    }
-
-    val nextNoiseScale = when {
-      noiseScale == null -> null
-      noiseScale.isNaN() -> state.noiseScale
-      else -> noiseScale
-    }
-    val nextNoiseScaleW = when {
-      noiseScaleW == null -> null
-      noiseScaleW.isNaN() -> state.noiseScaleW
-      else -> noiseScaleW
-    }
-    val nextLengthScale = when {
-      lengthScale == null -> null
-      lengthScale.isNaN() -> state.lengthScale
-      else -> lengthScale
-    }
-    try {
-      val result = detectTtsModel(state.modelDir, state.modelType)
-      if (result == null || result["success"] as? Boolean != true) {
-        Log.e("SherpaOnnxTts", "TTS_UPDATE_ERROR: Failed to re-detect TTS model")
-        promise.reject("TTS_UPDATE_ERROR", "Failed to re-detect TTS model")
-        return
-      }
-      val paths = (result["paths"] as? Map<*, *>)?.mapValues { (_, v) -> (v as? String).orEmpty() }?.mapKeys { it.key.toString() } ?: emptyMap()
-      val modelTypeStr = result["modelType"] as? String ?: state.modelType
-      val detectedModels = result["detectedModels"] as? ArrayList<*>
-
-      inst.tts?.release()
-      inst.tts = null
-      val config = buildTtsConfig(
-        paths, modelTypeStr, state.numThreads, state.debug,
-        nextNoiseScale, nextNoiseScaleW, nextLengthScale,
-        state.ruleFsts, state.ruleFars, state.maxNumSentences, state.silenceScale,
-        state.provider
-      )
-      inst.tts = OfflineTts(config = config)
-      val ttsInstance = inst.tts!!
-
-      val modelsArray = Arguments.createArray()
-      detectedModels?.forEach { modelObj ->
-        if (modelObj is HashMap<*, *>) {
-          val modelMap = Arguments.createMap()
-          modelMap.putString("type", modelObj["type"] as? String ?: "")
-          modelMap.putString("modelDir", modelObj["modelDir"] as? String ?: "")
-          modelsArray.pushMap(modelMap)
-        }
-      }
-
-      inst.ttsInitState = state.copy(
-        noiseScale = nextNoiseScale,
-        noiseScaleW = nextNoiseScaleW,
-        lengthScale = nextLengthScale
-      )
-
-      val resultMap = Arguments.createMap()
-      resultMap.putBoolean("success", true)
-      resultMap.putArray("detectedModels", modelsArray)
-      resultMap.putInt("sampleRate", ttsInstance.sampleRate())
-      resultMap.putInt("numSpeakers", ttsInstance.numSpeakers())
-      promise.resolve(resultMap)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_UPDATE_ERROR: Failed to update TTS params", e)
-      promise.reject("TTS_UPDATE_ERROR", "Failed to update TTS params", e)
-    }
-  }
-
-  fun generateTts(instanceId: String, text: String, options: ReadableMap?, promise: Promise) {
-    try {
-      val inst = getInstance(instanceId) ?: run {
-        Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS instance not found: $instanceId")
-        promise.reject("TTS_GENERATE_ERROR", "TTS instance not found: $instanceId")
-        return
-      }
-      if (!inst.hasEngine()) {
-        Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS not initialized")
-        promise.reject("TTS_GENERATE_ERROR", "TTS not initialized")
-        return
-      }
-      val sid = getSid(options)
-      val speed = getSpeed(options)
-      val audio = when {
-        hasReferenceAudio(options) && (inst.isZipvoice || inst.isPocket) -> {
-          if (inst.isZipvoice) {
-            val promptText = options!!.getString("referenceText")?.trim().orEmpty()
-            if (promptText.isEmpty()) {
-              Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Zipvoice voice cloning requires non-empty referenceText")
-              promise.reject(
-                "TTS_GENERATE_ERROR",
-                "Zipvoice voice cloning requires non-empty referenceText (transcript of reference audio)."
-              )
-              return
-            }
-          }
-          val config = parseGenerationConfig(options) ?: GenerationConfig(speed = speed, sid = sid)
-          inst.tts!!.generateWithConfig(text, config)
-        }
-        hasReferenceAudio(options) -> {
-          Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Reference audio is not supported for this TTS model type")
-          promise.reject(
-            "TTS_GENERATE_ERROR",
-            "Reference audio is only supported for Zipvoice and Pocket TTS."
-          )
-          return
-        }
-        inst.isPocket -> {
-          Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Pocket TTS requires reference audio for voice cloning")
-          promise.reject(
-            "TTS_GENERATE_ERROR",
-            "Pocket TTS requires reference audio for voice cloning. Pass referenceAudio and referenceSampleRate (> 0) in options."
-          )
-          return
-        }
-        else -> dispatchGenerate(inst, text, sid, speed)
-          ?: run {
-            Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS not initialized")
-            promise.reject("TTS_GENERATE_ERROR", "TTS not initialized")
-            return
-          }
-      }
-      val map = Arguments.createMap()
-      val samplesArray = Arguments.createArray()
-      for (sample in audio.samples) {
-        samplesArray.pushDouble(sample.toDouble())
-      }
-      map.putArray("samples", samplesArray)
-      map.putInt("sampleRate", audio.sampleRate)
-      promise.resolve(map)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "generateTts error: ${e.message}", e)
-      promise.reject("TTS_GENERATE_ERROR", e.message ?: "Failed to generate speech", e)
-    }
-  }
-
-  fun generateTtsWithTimestamps(instanceId: String, text: String, options: ReadableMap?, promise: Promise) {
-    try {
-      val inst = getInstance(instanceId) ?: run {
-        Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS instance not found: $instanceId")
-        promise.reject("TTS_GENERATE_ERROR", "TTS instance not found: $instanceId")
-        return
-      }
-      if (!inst.hasEngine()) {
-        Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS not initialized")
-        promise.reject("TTS_GENERATE_ERROR", "TTS not initialized")
-        return
-      }
-
-      val subtitleMode = getSubtitleMode(options)
-      val subtitleGranularity = getSubtitleGranularity(options)
-      if (isCharacterGranularityRequested(options) && subtitleMode != "accurate") {
-        Log.e(
-          "SherpaOnnxTts",
-          "TTS_SUBTITLE_ERROR: Character granularity is only supported when subtitleMode is 'accurate'"
-        )
-        promise.reject(
-          "TTS_SUBTITLE_ERROR",
-          "Character granularity is only supported when subtitleMode is 'accurate'."
-        )
-        return
-      }
-
-      val sid = getSid(options)
-      val speed = getSpeed(options)
-      val sentenceChunkSizes = mutableListOf<Int>()
-      val audio = when {
-        subtitleMode == "off" -> {
-          when {
-            hasReferenceAudio(options) && (inst.isZipvoice || inst.isPocket) -> {
-              if (inst.isZipvoice) {
-                val promptText = options!!.getString("referenceText")?.trim().orEmpty()
-                if (promptText.isEmpty()) {
-                  Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Zipvoice voice cloning requires non-empty referenceText")
-                  promise.reject(
-                    "TTS_GENERATE_ERROR",
-                    "Zipvoice voice cloning requires non-empty referenceText (transcript of reference audio)."
-                  )
-                  return
-                }
-              }
-              val config = parseGenerationConfig(options) ?: GenerationConfig(speed = speed, sid = sid)
-              inst.tts!!.generateWithConfig(text, config)
-            }
-            hasReferenceAudio(options) -> {
-              Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Reference audio is not supported for this TTS model type")
-              promise.reject(
-                "TTS_GENERATE_ERROR",
-                "Reference audio is only supported for Zipvoice and Pocket TTS."
-              )
-              return
-            }
-            inst.isPocket -> {
-              Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Pocket TTS requires reference audio for voice cloning")
-              promise.reject(
-                "TTS_GENERATE_ERROR",
-                "Pocket TTS requires reference audio for voice cloning. Pass referenceAudio and referenceSampleRate (> 0) in options."
-              )
-              return
-            }
-            else -> dispatchGenerate(inst, text, sid, speed)
-              ?: run {
-                Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS not initialized")
-                promise.reject("TTS_GENERATE_ERROR", "TTS not initialized")
-                return
-              }
-          }
-        }
-        hasReferenceAudio(options) && (inst.isZipvoice || inst.isPocket) -> {
-          if (inst.isZipvoice) {
-            val promptText = options!!.getString("referenceText")?.trim().orEmpty()
-            if (promptText.isEmpty()) {
-              Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Zipvoice voice cloning requires non-empty referenceText")
-              promise.reject(
-                "TTS_GENERATE_ERROR",
-                "Zipvoice voice cloning requires non-empty referenceText (transcript of reference audio)."
-              )
-              return
-            }
-          }
-          val config = parseGenerationConfig(options) ?: GenerationConfig(speed = speed, sid = sid)
-          inst.tts!!.generateWithConfigAndCallback(
-            text,
-            config,
-            ttsChunkCallbackForJni(sentenceChunkSizes)
-          )
-        }
-        hasReferenceAudio(options) -> {
-          Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Reference audio is not supported for this TTS model type")
-          promise.reject(
-            "TTS_GENERATE_ERROR",
-            "Reference audio is only supported for Zipvoice and Pocket TTS."
-          )
-          return
-        }
-        inst.isPocket -> {
-          Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: Pocket TTS requires reference audio for voice cloning")
-          promise.reject(
-            "TTS_GENERATE_ERROR",
-            "Pocket TTS requires reference audio for voice cloning. Pass referenceAudio and referenceSampleRate (> 0) in options."
-          )
-          return
-        }
-        else -> {
-          val tts = inst.tts
-          if (tts == null) {
-            Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: TTS not initialized")
-            promise.reject("TTS_GENERATE_ERROR", "TTS not initialized")
-            return
-          }
-          tts.generateWithCallback(text, sid, speed, ttsChunkCallbackForJni(sentenceChunkSizes))
-        }
-      }
-
-      if (subtitleMode != "off" && sentenceChunkSizes.isEmpty() && audio.samples.isNotEmpty()) {
-        sentenceChunkSizes.add(audio.samples.size)
-      }
-
-      val map = Arguments.createMap()
-      val samplesArray = Arguments.createArray()
-      for (sample in audio.samples) {
-        samplesArray.pushDouble(sample.toDouble())
-      }
-      map.putArray("samples", samplesArray)
-      map.putInt("sampleRate", audio.sampleRate)
-
-      val subtitleItems = if (subtitleMode == "off") {
-        emptyList()
-      } else {
-        val sentenceSegments = SherpaOnnxTextSegmenter.splitIntoSentences(text)
-        if (subtitleGranularity == "word") {
-          SherpaOnnxTextSegmenter.buildWordSubtitlesFromSentenceChunks(
-            sentenceSegments,
-            sentenceChunkSizes,
-            audio.sampleRate
-          )
-        } else {
-          SherpaOnnxTextSegmenter.buildSubtitlesFromChunks(
-            sentenceSegments,
-            sentenceChunkSizes,
-            audio.sampleRate
-          )
-        }
-      }
-
-      map.putArray("subtitles", toSubtitleWritableArray(subtitleItems))
-      val timingMode = if (subtitleMode == "off") "off" else "estimated"
-      map.putString("timingMode", timingMode)
-      promise.resolve(map)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_GENERATE_ERROR: ${e.message ?: "Failed to generate speech"}", e)
-      promise.reject("TTS_GENERATE_ERROR", e.message ?: "Failed to generate speech", e)
-    }
-  }
-
-  fun generateTtsStream(instanceId: String, requestId: String, text: String, options: ReadableMap?, promise: Promise) {
-    val inst = getInstance(instanceId) ?: run {
-      Log.e("SherpaOnnxTts", "TTS_STREAM_ERROR: TTS instance not found: $instanceId")
-      promise.reject("TTS_STREAM_ERROR", "TTS instance not found: $instanceId")
-      return
-    }
-    if (inst.ttsStreamRunning.get()) {
-      Log.e("SherpaOnnxTts", "TTS_STREAM_ERROR: TTS streaming already in progress")
-      promise.reject("TTS_STREAM_ERROR", "TTS streaming already in progress")
-      return
-    }
-    if (!inst.hasEngine()) {
-      Log.e("SherpaOnnxTts", "TTS_STREAM_ERROR: TTS not initialized")
-      promise.reject("TTS_STREAM_ERROR", "TTS not initialized")
-      return
-    }
-    if (inst.isPocket && !hasReferenceAudio(options)) {
-      Log.e("SherpaOnnxTts", "TTS_STREAM_ERROR: Pocket TTS requires reference audio for voice cloning")
-      promise.reject(
-        "TTS_STREAM_ERROR",
-        "Pocket TTS requires reference audio for voice cloning. Pass referenceAudio and referenceSampleRate (> 0) in options."
-      )
-      return
-    }
-    if (hasReferenceAudio(options) && inst.isZipvoice) {
-      Log.e("SherpaOnnxTts", "TTS_STREAM_ERROR: Streaming with reference audio not supported for Zipvoice")
-      promise.reject("TTS_STREAM_ERROR", "Streaming with reference audio not supported for Zipvoice")
-      return
-    }
-    if (hasReferenceAudio(options) && !inst.isPocket) {
-      Log.e("SherpaOnnxTts", "TTS_STREAM_ERROR: Reference audio streaming is only supported for Pocket TTS")
-      promise.reject(
-        "TTS_STREAM_ERROR",
-        "Reference audio streaming is only supported for Pocket TTS."
-      )
-      return
-    }
-    val sid = getSid(options)
-    val speed = getSpeed(options)
-    inst.ttsStreamCancelled.set(false)
-    inst.ttsStreamRunning.set(true)
-    inst.ttsStreamThread = Thread {
-      try {
-        val sampleRate = dispatchSampleRate(inst)
-        when {
-          hasReferenceAudio(options) && inst.isPocket -> {
-            val config = parseGenerationConfig(options) ?: GenerationConfig(speed = speed, sid = sid)
-            inst.tts!!.generateWithConfigAndCallback(
-              text,
-              config,
-              ttsStreamChunkCallbackForJni(inst.ttsStreamCancelled) { chunk ->
-                emitChunk(instanceId, requestId, chunk, sampleRate, 0f, false)
-              }
-            )
-          }
-          else -> {
-            inst.tts!!.generateWithCallback(
-              text,
-              sid,
-              speed,
-              ttsStreamChunkCallbackForJni(inst.ttsStreamCancelled) { chunk ->
-                emitChunk(instanceId, requestId, chunk, sampleRate, 0f, false)
-              }
-            )
-          }
-        }
-        if (!inst.ttsStreamCancelled.get()) {
-          emitChunk(instanceId, requestId, FloatArray(0), sampleRate, 1f, true)
-        }
-      } catch (e: Exception) {
-        if (!inst.ttsStreamCancelled.get()) {
-          emitError(instanceId, requestId, "TTS streaming failed: ${e.message}")
-        }
-      } finally {
-        emitEnd(instanceId, requestId, inst.ttsStreamCancelled.get())
-        inst.ttsStreamRunning.set(false)
-      }
-    }
-    inst.ttsStreamThread?.start()
-    promise.resolve(null)
-  }
-
-  fun cancelTtsStream(instanceId: String, promise: Promise) {
-    val inst = getInstance(instanceId)
-    if (inst != null) {
-      inst.ttsStreamCancelled.set(true)
-      inst.ttsStreamThread?.interrupt()
-    }
-    promise.resolve(null)
-  }
-
-  fun startTtsPcmPlayer(instanceId: String, sampleRate: Double, channels: Double, promise: Promise) {
-    val inst = getInstance(instanceId) ?: run {
-      Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: TTS instance not found: $instanceId")
-      promise.reject("TTS_PCM_ERROR", "TTS instance not found: $instanceId")
-      return
-    }
-    try {
-      if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-        Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: PCM playback requires API 21+")
-        promise.reject("TTS_PCM_ERROR", "PCM playback requires API 21+")
-        return
-      }
-      if (channels.toInt() != 1) {
-        Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: PCM playback supports mono only")
-        promise.reject("TTS_PCM_ERROR", "PCM playback supports mono only")
-        return
-      }
-      inst.stopPcmPlayer()
-      val channelConfig = AudioFormat.CHANNEL_OUT_MONO
-      val audioFormat = AudioFormat.Builder()
-        .setSampleRate(sampleRate.toInt())
-        .setChannelMask(channelConfig)
-        .setEncoding(AudioFormat.ENCODING_PCM_FLOAT)
-        .build()
-      val minBufferSize = AudioTrack.getMinBufferSize(sampleRate.toInt(), channelConfig, AudioFormat.ENCODING_PCM_FLOAT)
-      if (minBufferSize == AudioTrack.ERROR || minBufferSize == AudioTrack.ERROR_BAD_VALUE) {
-        Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: Invalid buffer size for PCM player")
-        promise.reject("TTS_PCM_ERROR", "Invalid buffer size for PCM player")
-        return
-      }
-      val attributes = AudioAttributes.Builder()
-        .setUsage(AudioAttributes.USAGE_MEDIA)
-        .setContentType(AudioAttributes.CONTENT_TYPE_SPEECH)
-        .build()
-      inst.ttsPcmTrack = AudioTrack(attributes, audioFormat, minBufferSize, AudioTrack.MODE_STREAM, AudioManager.AUDIO_SESSION_ID_GENERATE)
-      inst.ttsPcmTrack?.play()
-      promise.resolve(null)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: Failed to start PCM player", e)
-      promise.reject("TTS_PCM_ERROR", "Failed to start PCM player", e)
-    }
-  }
-
-  fun writeTtsPcmChunk(instanceId: String, samples: ReadableArray, promise: Promise) {
-    val inst = getInstance(instanceId) ?: run {
-      Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: TTS instance not found: $instanceId")
-      promise.reject("TTS_PCM_ERROR", "TTS instance not found: $instanceId")
-      return
-    }
-    val track = inst.ttsPcmTrack ?: run {
-      Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: PCM player not initialized")
-      promise.reject("TTS_PCM_ERROR", "PCM player not initialized")
-      return
-    }
-    try {
-      val buffer = FloatArray(samples.size())
-      for (i in 0 until samples.size()) {
-        buffer[i] = samples.getDouble(i).toFloat()
-      }
-      val written = track.write(buffer, 0, buffer.size, AudioTrack.WRITE_BLOCKING)
-      if (written < 0) {
-        Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: PCM write failed: $written")
-        promise.reject("TTS_PCM_ERROR", "PCM write failed: $written")
-        return
-      }
-      promise.resolve(null)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: Failed to write PCM chunk", e)
-      promise.reject("TTS_PCM_ERROR", "Failed to write PCM chunk", e)
-    }
-  }
-
-  fun stopTtsPcmPlayer(instanceId: String, promise: Promise) {
-    try {
-      getInstance(instanceId)?.stopPcmPlayer()
-      promise.resolve(null)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_PCM_ERROR: Failed to stop PCM player", e)
-      promise.reject("TTS_PCM_ERROR", "Failed to stop PCM player", e)
-    }
-  }
-
-  fun getTtsSampleRate(instanceId: String, promise: Promise) {
-    try {
-      val inst = getInstance(instanceId) ?: run {
-        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS instance not found: $instanceId")
-        promise.reject("TTS_ERROR", "TTS instance not found: $instanceId")
-        return
-      }
-      if (!inst.hasEngine()) {
-        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS not initialized")
-        promise.reject("TTS_ERROR", "TTS not initialized")
-        return
-      }
-      promise.resolve(dispatchSampleRate(inst).toDouble())
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_ERROR: Failed to get sample rate", e)
-      promise.reject("TTS_ERROR", "Failed to get sample rate", e)
-    }
-  }
-
-  fun getTtsNumSpeakers(instanceId: String, promise: Promise) {
-    try {
-      val inst = getInstance(instanceId) ?: run {
-        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS instance not found: $instanceId")
-        promise.reject("TTS_ERROR", "TTS instance not found: $instanceId")
-        return
-      }
-      if (!inst.hasEngine()) {
-        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS not initialized")
-        promise.reject("TTS_ERROR", "TTS not initialized")
-        return
-      }
-      promise.resolve(dispatchNumSpeakers(inst).toDouble())
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_ERROR: Failed to get number of speakers", e)
-      promise.reject("TTS_ERROR", "Failed to get number of speakers", e)
-    }
-  }
-
-  fun unloadTts(instanceId: String, promise: Promise) {
-    try {
-      val inst = instances.remove(instanceId)
-      if (inst != null) {
-        inst.stopPcmPlayer()
-        inst.releaseEngines()
-      }
-      promise.resolve(null)
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_RELEASE_ERROR: Failed to release TTS resources", e)
-      promise.reject("TTS_RELEASE_ERROR", "Failed to release TTS resources", e)
-    }
-  }
-
-  fun saveTtsAudio(
-    samples: ReadableArray,
-    sampleRate: Double,
-    destinationType: String,
-    pathOrDirectoryUri: String,
-    filename: String,
-    format: String,
-    outputSampleRateHz: Double,
-    promise: Promise
-  ) {
-    try {
-      val fmt = format.trim().lowercase().ifEmpty { "wav" }
-      val outHz = outputSampleRateHz.toInt()
-      val rateErr = validateTtsEncodeOutputSampleRate(fmt, outHz)
-      if (rateErr != null) {
-        promise.reject("TTS_SAVE_ERROR", rateErr)
-        return
-      }
-
-      when (destinationType) {
-        "file" -> {
-          if (fmt == "wav" || fmt == "wav16k") {
-            saveTtsAudioToFile(samples, sampleRate, pathOrDirectoryUri, promise)
-            return
-          }
-          encodeToFileThenResolve(samples, sampleRate, pathOrDirectoryUri, fmt, outHz, promise)
-        }
-        "androidContent" -> {
-          if (filename.isBlank()) {
-            promise.reject("TTS_SAVE_ERROR", "filename is required for androidContent destination")
-            return
-          }
-          if (fmt == "wav" || fmt == "wav16k") {
-            saveTtsAudioToContentUri(samples, sampleRate, pathOrDirectoryUri, filename, promise)
-            return
-          }
-          val ext = fileExtensionForFormat(fmt)
-          val cacheOut = File(context.cacheDir, "tts_save_${System.nanoTime()}.$ext")
-          encodeToFileThenCopyToContentUri(
-            samples,
-            sampleRate,
-            cacheOut.absolutePath,
-            fmt,
-            outHz,
-            pathOrDirectoryUri,
-            filename,
-            mimeTypeForTtsFormat(fmt),
-            promise
-          )
-        }
-        else -> promise.reject("TTS_SAVE_ERROR", "Invalid destinationType: use 'file' or 'androidContent'")
-      }
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_SAVE_ERROR: saveTtsAudio failed", e)
-      promise.reject("TTS_SAVE_ERROR", "Failed to save TTS audio: ${e.message}", e)
-    }
-  }
-
-  private fun validateTtsEncodeOutputSampleRate(format: String, rate: Int): String? {
-    if (rate < 0) return "Invalid outputSampleRateHz: must be >= 0"
-    if (format == "mp3") {
-      val allowed = setOf(0, 32000, 44100, 48000)
-      if (!allowed.contains(rate)) {
-        return "MP3 output sample rate must be one of 32000, 44100, 48000, or 0 (default). Received: $rate"
-      }
-    } else if (format == "opus" || format == "oggm" || format == "webm" || format == "mkv" || format == "ogg") {
-      val allowed = setOf(0, 8000, 12000, 16000, 24000, 48000)
-      if (!allowed.contains(rate)) {
-        return "Opus output sample rate must be 8000, 12000, 16000, 24000, 48000, or 0 (default). Received: $rate"
-      }
-    }
-    return null
-  }
-
-  private fun fileExtensionForFormat(format: String): String = when (format) {
-    "mp3" -> "mp3"
-    "flac" -> "flac"
-    "m4a", "aac" -> "m4a"
-    "opus" -> "opus"
-    "ogg", "oggm" -> "ogg"
-    "webm" -> "webm"
-    "mkv" -> "mkv"
-    else -> "bin"
-  }
-
-  private fun mimeTypeForTtsFormat(format: String): String = when (format) {
-    "mp3" -> "audio/mpeg"
-    "flac" -> "audio/flac"
-    "m4a", "aac" -> "audio/mp4"
-    "opus", "ogg", "oggm" -> "audio/ogg"
-    "webm" -> "audio/webm"
-    "wav", "wav16k" -> "audio/wav"
-    else -> "application/octet-stream"
-  }
-
-  private fun readableArrayToFloatArray(samples: ReadableArray): FloatArray {
-    val n = samples.size()
-    val out = FloatArray(n)
-    for (i in 0 until n) {
-      out[i] = samples.getDouble(i).toFloat()
-    }
-    return out
-  }
-
-  private fun writeMonoF32LeRawFile(floats: FloatArray, file: File) {
-    val bb = ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN)
-    FileOutputStream(file).use { o ->
-      for (f in floats) {
-        bb.clear()
-        bb.putFloat(f)
-        o.write(bb.array())
-      }
-    }
-  }
-
-  private fun encodeToFileThenResolve(
-    samples: ReadableArray,
-    sampleRate: Double,
-    outputPath: String,
-    format: String,
-    outputSampleRateHz: Int,
-    promise: Promise
-  ) {
-    val floats = readableArrayToFloatArray(samples)
-    val raw = File(context.cacheDir, "tts_pcm_${System.nanoTime()}.raw")
-    try {
-      writeMonoF32LeRawFile(floats, raw)
-      val err = encodeMonoFromRawFile(raw.absolutePath, sampleRate.toInt(), outputPath, format, outputSampleRateHz)
-      if (err.isNotEmpty()) {
-        promise.reject("TTS_SAVE_ERROR", err)
-        return
-      }
-      promise.resolve(outputPath)
-    } finally {
-      raw.delete()
-    }
-  }
-
-  private fun encodeToFileThenCopyToContentUri(
-    samples: ReadableArray,
-    sampleRate: Double,
-    cacheOutputPath: String,
-    format: String,
-    outputSampleRateHz: Int,
-    directoryUri: String,
-    filename: String,
-    mimeType: String,
-    promise: Promise
-  ) {
-    val floats = readableArrayToFloatArray(samples)
-    val raw = File(context.cacheDir, "tts_pcm_${System.nanoTime()}.raw")
-    val cacheFile = File(cacheOutputPath)
-    try {
-      writeMonoF32LeRawFile(floats, raw)
-      val err = encodeMonoFromRawFile(raw.absolutePath, sampleRate.toInt(), cacheOutputPath, format, outputSampleRateHz)
-      if (err.isNotEmpty()) {
-        promise.reject("TTS_SAVE_ERROR", err)
-        return
-      }
-      if (!cacheFile.isFile || !cacheFile.canRead()) {
-        promise.reject("TTS_SAVE_ERROR", "Encoded file missing: $cacheOutputPath")
-        return
-      }
-      val resolver = context.contentResolver
-      val dirUri = Uri.parse(directoryUri)
-      val fileUri = SherpaOnnxContentUriUtils.createDocumentInDirectory(resolver, dirUri, filename, mimeType)
-      FileInputStream(cacheFile).use { inputStream ->
-        resolver.openOutputStream(fileUri, "w")?.use { outputStream ->
-          inputStream.copyTo(outputStream)
-          outputStream.flush()
-        } ?: throw IllegalStateException("Failed to open output stream for URI: $fileUri")
-      }
-      promise.resolve(fileUri.toString())
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_SAVE_ERROR: encode+SAF copy failed", e)
-      promise.reject("TTS_SAVE_ERROR", "Failed to save encoded audio to content URI: ${e.message}", e)
-    } finally {
-      raw.delete()
-      cacheFile.delete()
-    }
-  }
-
-  private fun saveTtsAudioToFile(
-    samples: ReadableArray,
-    sampleRate: Double,
-    filePath: String,
-    promise: Promise
-  ) {
-    try {
-      val samplesArray = FloatArray(samples.size())
-      for (i in 0 until samples.size()) {
-        samplesArray[i] = samples.getDouble(i).toFloat()
-      }
-      val success = GeneratedAudio(samplesArray, sampleRate.toInt()).save(filePath)
-      if (success) {
-        promise.resolve(filePath)
-      } else {
-        Log.e("SherpaOnnxTts", "TTS_SAVE_ERROR: Failed to save audio to file")
-        promise.reject("TTS_SAVE_ERROR", "Failed to save audio to file")
-      }
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_SAVE_ERROR: Failed to save audio to file", e)
-      promise.reject("TTS_SAVE_ERROR", "Failed to save audio to file", e)
-    }
-  }
-
-  private fun saveTtsAudioToContentUri(
-    samples: ReadableArray,
-    sampleRate: Double,
-    directoryUri: String,
-    filename: String,
-    promise: Promise
-  ) {
-    try {
-      val samplesArray = FloatArray(samples.size())
-      for (i in 0 until samples.size()) {
-        samplesArray[i] = samples.getDouble(i).toFloat()
-      }
-      val resolver = context.contentResolver
-      val dirUri = Uri.parse(directoryUri)
-      val fileUri = SherpaOnnxContentUriUtils.createDocumentInDirectory(resolver, dirUri, filename, "audio/wav")
-      resolver.openOutputStream(fileUri, "w")?.use { outputStream ->
-        writeWavToStream(samplesArray, sampleRate.toInt(), outputStream)
-      } ?: throw IllegalStateException("Failed to open output stream for URI: $fileUri")
-      promise.resolve(fileUri.toString())
-    } catch (e: Exception) {
-      Log.e("SherpaOnnxTts", "TTS_SAVE_ERROR: Failed to save audio to content URI", e)
-      promise.reject("TTS_SAVE_ERROR", "Failed to save audio to content URI", e)
-    }
-  }
-
-  // -- Dual-engine dispatch helpers --
-
-  /**
-   * True when voice-cloning reference audio is present and valid for native use:
-   * non-empty [referenceAudio] array and [referenceSampleRate] > 0.
-   * [referenceText] alone does not enable cloning (matches sherpa-onnx behavior).
-   */
-  private fun hasReferenceAudio(options: ReadableMap?): Boolean {
-    if (options == null) return false
-    val refAudio = options.getArray("referenceAudio") ?: return false
-    if (refAudio.size() == 0) return false
-    return readReferenceSampleRate(options) > 0
-  }
-
-  private fun readReferenceSampleRate(options: ReadableMap): Int =
-    if (options.hasKey("referenceSampleRate")) options.getDouble("referenceSampleRate").toInt() else 0
-
-  /** Parse sid and speed from options with defaults. */
-  private fun getSid(options: ReadableMap?): Int =
-    if (options != null && options.hasKey("sid")) options.getDouble("sid").toInt() else 0
-
-  private fun getSpeed(options: ReadableMap?): Float =
-    if (options != null && options.hasKey("speed")) options.getDouble("speed").toFloat() else 1.0f
-
-  private fun getSubtitleMode(options: ReadableMap?): String {
-    val raw = options?.getString("subtitleMode")?.trim()?.lowercase()
-    return when (raw) {
-      "off", "fast", "accurate" -> raw
-      else -> "fast"
-    }
-  }
-
-  private fun getSubtitleGranularity(options: ReadableMap?): String {
-    val raw = options?.getString("subtitleGranularity")?.trim()?.lowercase()
-    return when (raw) {
-      "word", "sentence" -> raw
-      else -> "sentence"
-    }
-  }
-
-  private fun isCharacterGranularityRequested(options: ReadableMap?): Boolean {
-    val raw = options?.getString("subtitleGranularity")?.trim()?.lowercase()
-    return raw == "character"
-  }
-
-  private fun toSubtitleWritableArray(items: List<SubtitleTimingItem>) = Arguments.createArray().apply {
-    for (item in items) {
-      val subtitleMap = Arguments.createMap()
-      subtitleMap.putString("text", item.text)
-      subtitleMap.putDouble("start", item.start)
-      subtitleMap.putDouble("end", item.end)
-      pushMap(subtitleMap)
-    }
-  }
-
-  /** Build Kotlin GenerationConfig from ReadableMap. Returns null only when options is null; otherwise returns a config with sid, speed, silenceScale, numSteps, and any reference/extra fields from options. */
-  private fun parseGenerationConfig(options: ReadableMap?): GenerationConfig? {
-    if (options == null) return null
-    val refAudio = options.getArray("referenceAudio")
-    val refSampleRate = if (options.hasKey("referenceSampleRate")) options.getDouble("referenceSampleRate").toInt() else 0
-    val refText = options.getString("referenceText")
-    val silenceScale = if (options.hasKey("silenceScale")) options.getDouble("silenceScale").toFloat() else 0.2f
-    val speed = getSpeed(options)
-    val sid = getSid(options)
-    val numSteps = if (options.hasKey("numSteps")) options.getDouble("numSteps").toInt() else 5
-    val extraMap = options.getMap("extra")?.let { map ->
-      val it = map.keySetIterator()
-      buildMap<String, String> {
-        while (it.hasNextKey()) {
-          val k = it.nextKey()
-          put(k, map.getString(k).orEmpty())
-        }
-      }
-    }
-    val refAudioFloat = refAudio?.let { arr ->
-      FloatArray(arr.size()) { i -> arr.getDouble(i).toFloat() }
-    }
-    return GenerationConfig(
-      silenceScale = silenceScale,
-      speed = speed,
-      sid = sid,
-      referenceAudio = refAudioFloat,
-      referenceSampleRate = refSampleRate,
-      referenceText = refText,
-      numSteps = numSteps,
-      extra = extraMap
-    )
-  }
-
-  /** Dispatch generate to whichever engine is active on the instance. Returns null if none loaded. */
-  private fun dispatchGenerate(inst: TtsEngineInstance, text: String, sid: Int, speed: Float): GeneratedAudio? {
-    return inst.tts?.generate(text, sid, speed)
-  }
-
-  private fun dispatchSampleRate(inst: TtsEngineInstance): Int {
-    return inst.tts?.sampleRate() ?: 0
-  }
-
-  private fun dispatchNumSpeakers(inst: TtsEngineInstance): Int {
-    return inst.tts?.numSpeakers() ?: 0
-  }
-
-  private fun path(paths: Map<String, String>, key: String): String = paths[key].orEmpty()
-
-  private fun buildTtsConfig(
-    paths: Map<String, String>,
-    modelType: String,
-    numThreads: Int,
-    debug: Boolean,
-    noiseScale: Double?,
-    noiseScaleW: Double?,
-    lengthScale: Double?,
-    ruleFsts: String?,
-    ruleFars: String?,
-    maxNumSentences: Int?,
-    silenceScale: Double?,
-    provider: String?
-  ): OfflineTtsConfig {
-    val ns = noiseScale?.toFloat() ?: 0.667f
-    val nsw = noiseScaleW?.toFloat() ?: 0.8f
-    val ls = lengthScale?.toFloat() ?: 1.0f
-    val prov = provider?.takeIf { it.isNotBlank() } ?: "cpu"
-    val modelConfig = when (modelType) {
-      "vits" -> OfflineTtsModelConfig(
-        vits = OfflineTtsVitsModelConfig(
-          model = path(paths, "ttsModel"),
-          lexicon = path(paths, "lexicon"),
-          tokens = path(paths, "tokens"),
-          dataDir = path(paths, "dataDir"),
-          noiseScale = ns,
-          noiseScaleW = nsw,
-          lengthScale = ls
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      "matcha" -> OfflineTtsModelConfig(
-        matcha = OfflineTtsMatchaModelConfig(
-          acousticModel = path(paths, "acousticModel"),
-          vocoder = path(paths, "vocoder"),
-          lexicon = path(paths, "lexicon"),
-          tokens = path(paths, "tokens"),
-          dataDir = path(paths, "dataDir"),
-          noiseScale = ns,
-          lengthScale = ls
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      "kokoro" -> OfflineTtsModelConfig(
-        kokoro = OfflineTtsKokoroModelConfig(
-          model = path(paths, "ttsModel"),
-          voices = path(paths, "voices"),
-          tokens = path(paths, "tokens"),
-          dataDir = path(paths, "dataDir"),
-          lexicon = path(paths, "lexicon"),
-          lengthScale = ls
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      "kitten" -> OfflineTtsModelConfig(
-        kitten = OfflineTtsKittenModelConfig(
-          model = path(paths, "ttsModel"),
-          voices = path(paths, "voices"),
-          tokens = path(paths, "tokens"),
-          dataDir = path(paths, "dataDir"),
-          lengthScale = ls
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      "pocket" -> OfflineTtsModelConfig(
-        pocket = OfflineTtsPocketModelConfig(
-          lmFlow = path(paths, "lmFlow"),
-          lmMain = path(paths, "lmMain"),
-          encoder = path(paths, "encoder"),
-          decoder = path(paths, "decoder"),
-          textConditioner = path(paths, "textConditioner"),
-          vocabJson = path(paths, "vocabJson"),
-          tokenScoresJson = path(paths, "tokenScoresJson")
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      "zipvoice" -> OfflineTtsModelConfig(
-        zipvoice = OfflineTtsZipVoiceModelConfig(
-          tokens = path(paths, "tokens"),
-          encoder = path(paths, "encoder"),
-          decoder = path(paths, "decoder"),
-          vocoder = path(paths, "vocoder"),
-          dataDir = path(paths, "dataDir"),
-          lexicon = path(paths, "lexicon")
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      "supertonic" -> OfflineTtsModelConfig(
-        supertonic = OfflineTtsSupertonicModelConfig(
-          durationPredictor = path(paths, "durationPredictor"),
-          textEncoder = path(paths, "textEncoder"),
-          vectorEstimator = path(paths, "vectorEstimator"),
-          vocoder = path(paths, "vocoder"),
-          ttsJson = path(paths, "ttsJson"),
-          unicodeIndexer = path(paths, "unicodeIndexer"),
-          voiceStyle = path(paths, "voiceStyle")
-        ),
-        numThreads = numThreads,
-        debug = debug,
-        provider = prov
-      )
-      else -> {
-        if (path(paths, "acousticModel").isNotEmpty()) {
-          OfflineTtsModelConfig(
-            matcha = OfflineTtsMatchaModelConfig(
-              acousticModel = path(paths, "acousticModel"),
-              vocoder = path(paths, "vocoder"),
-              lexicon = path(paths, "lexicon"),
-              tokens = path(paths, "tokens"),
-              dataDir = path(paths, "dataDir"),
-              noiseScale = ns,
-              lengthScale = ls
-            ),
-            numThreads = numThreads,
-            debug = debug,
-            provider = prov
-          )
-        } else if (path(paths, "voices").isNotEmpty()) {
-          OfflineTtsModelConfig(
-            kokoro = OfflineTtsKokoroModelConfig(
-              model = path(paths, "ttsModel"),
-              voices = path(paths, "voices"),
-              tokens = path(paths, "tokens"),
-              dataDir = path(paths, "dataDir"),
-              lexicon = path(paths, "lexicon"),
-              lengthScale = ls
-            ),
-            numThreads = numThreads,
-            debug = debug,
-            provider = prov
-          )
-        } else {
-          OfflineTtsModelConfig(
-            vits = OfflineTtsVitsModelConfig(
-              model = path(paths, "ttsModel"),
-              lexicon = path(paths, "lexicon"),
-              tokens = path(paths, "tokens"),
-              dataDir = path(paths, "dataDir"),
-              noiseScale = ns,
-              noiseScaleW = nsw,
-              lengthScale = ls
-            ),
-            numThreads = numThreads,
-            debug = debug,
-            provider = prov
-          )
-        }
-      }
-    }
-    return OfflineTtsConfig(
-      model = modelConfig,
-      ruleFsts = ruleFsts?.takeIf { it.isNotBlank() } ?: "",
-      ruleFars = ruleFars?.takeIf { it.isNotBlank() } ?: "",
-      maxNumSentences = maxNumSentences?.coerceAtLeast(1) ?: 1,
-      silenceScale = silenceScale?.toFloat()?.coerceIn(0f, 10f) ?: 0.2f
-    )
-  }
-
-  private fun writeWavToStream(samples: FloatArray, sampleRate: Int, outputStream: OutputStream) {
-    val numChannels = 1
-    val bitsPerSample = 16
-    val byteRate = sampleRate * numChannels * bitsPerSample / 8
-    val blockAlign = numChannels * bitsPerSample / 8
-    val dataSize = samples.size * 2
-    val chunkSize = 36 + dataSize
-    outputStream.write("RIFF".toByteArray(Charsets.US_ASCII))
-    writeIntLE(outputStream, chunkSize)
-    outputStream.write("WAVE".toByteArray(Charsets.US_ASCII))
-    outputStream.write("fmt ".toByteArray(Charsets.US_ASCII))
-    writeIntLE(outputStream, 16)
-    writeShortLE(outputStream, 1)
-    writeShortLE(outputStream, numChannels.toShort())
-    writeIntLE(outputStream, sampleRate)
-    writeIntLE(outputStream, byteRate)
-    writeShortLE(outputStream, blockAlign.toShort())
-    writeShortLE(outputStream, bitsPerSample.toShort())
-    outputStream.write("data".toByteArray(Charsets.US_ASCII))
-    writeIntLE(outputStream, dataSize)
-    for (sample in samples) {
-      val clamped = sample.coerceIn(-1.0f, 1.0f)
-      val intSample = (clamped * 32767.0f).toInt()
-      writeShortLE(outputStream, intSample.toShort())
-    }
-    outputStream.flush()
-  }
-
-  private fun writeIntLE(outputStream: OutputStream, value: Int) {
-    outputStream.write(value and 0xFF)
-    outputStream.write((value shr 8) and 0xFF)
-    outputStream.write((value shr 16) and 0xFF)
-    outputStream.write((value shr 24) and 0xFF)
-  }
-
-  private fun writeShortLE(outputStream: OutputStream, value: Short) {
-    val intValue = value.toInt()
-    outputStream.write(intValue and 0xFF)
-    outputStream.write((intValue shr 8) and 0xFF)
-  }
-
-}
