+++ android/src/main/java/com/sherpaonnx/tts/core/SherpaOnnxTtsHelper.kt
@@ -1,0 +1,4 @@
+package com.sherpaonnx.tts.core
+
+/** @suppress Back-compat name for [SherpaOnnxTtsCoordinator]. */
+internal typealias SherpaOnnxTtsHelper = SherpaOnnxTtsCoordinator
