+++ android/src/main/java/com/sherpaonnx/tts/facade/SherpaOnnxOfflineTtsHelper.kt
@@ -1,0 +1,55 @@
+package com.sherpaonnx.tts.facade
+
+import com.facebook.react.bridge.Promise
+import com.facebook.react.bridge.ReadableMap
+import com.sherpaonnx.tts.core.SherpaOnnxTtsHelper
+
+/**
+ * Offline/batch-facing TTS facade.
+ * Delegates to [SherpaOnnxTtsCoordinator] (see [SherpaOnnxTtsHelper] typealias).
+ */
+internal class SherpaOnnxOfflineTtsHelper(
+  private val core: SherpaOnnxTtsHelper
+) {
+  fun updateTtsParams(
+    instanceId: String,
+    noiseScale: Double?,
+    noiseScaleW: Double?,
+    lengthScale: Double?,
+    promise: Promise
+  ) = core.updateTtsParams(instanceId, noiseScale, noiseScaleW, lengthScale, promise)
+
+  fun generateTts(
+    instanceId: String,
+    text: String,
+    options: ReadableMap?,
+    promise: Promise
+  ) = core.generateTts(instanceId, text, options, promise)
+
+  fun generateTtsWithTimestamps(
+    instanceId: String,
+    text: String,
+    options: ReadableMap?,
+    promise: Promise
+  ) = core.generateTtsWithTimestamps(instanceId, text, options, promise)
+
+  fun saveTtsAudio(
+    samples: com.facebook.react.bridge.ReadableArray,
+    sampleRate: Double,
+    destinationType: String,
+    pathOrDirectoryUri: String,
+    filename: String,
+    format: String,
+    outputSampleRateHz: Double,
+    promise: Promise
+  ) = core.saveTtsAudio(
+    samples,
+    sampleRate,
+    destinationType,
+    pathOrDirectoryUri,
+    filename,
+    format,
+    outputSampleRateHz,
+    promise
+  )
+}
