+++ android/src/main/java/com/sherpaonnx/tts/service/TtsLifecycleService.kt
@@ -1,0 +1,182 @@
+package com.sherpaonnx.tts.service
+
+import android.util.Log
+import com.facebook.react.bridge.Arguments
+import com.facebook.react.bridge.Promise
+import com.facebook.react.bridge.ReadableArray
+import com.sherpaonnx.tts.core.TtsEngineRepository
+import com.sherpaonnx.tts.core.dispatchNumSpeakers
+import com.sherpaonnx.tts.core.dispatchSampleRate
+
+/**
+ * Engine lifecycle, metadata, catalog helpers, and init executor shutdown.
+ */
+internal class TtsLifecycleService(
+  private val repository: TtsEngineRepository,
+  private val ttsInitExecutor: java.util.concurrent.ExecutorService,
+  private val nativeDetectTtsModel: (String, String) -> HashMap<String, Any>?,
+  private val nativeBatchTtsCatalogHints: (Array<String>) -> java.util.ArrayList<java.util.HashMap<String, Any>>
+) {
+  /**
+   * Shuts down the TTS init executor and releases all engine instances.
+   * Call from the native module's onCatalystInstanceDestroy() to avoid leaking the executor thread.
+   */
+  fun shutdown() {
+    try {
+      ttsInitExecutor.shutdown()
+      if (!ttsInitExecutor.awaitTermination(3, java.util.concurrent.TimeUnit.SECONDS)) {
+        ttsInitExecutor.shutdownNow()
+      }
+    } catch (e: InterruptedException) {
+      Thread.currentThread().interrupt()
+      ttsInitExecutor.shutdownNow()
+    }
+    repository.forEachInstance { inst ->
+      inst.releaseEngines()
+      inst.stopPcmPlayer()
+    }
+    repository.clear()
+  }
+
+  fun getTtsSampleRate(instanceId: String, promise: Promise) {
+    try {
+      val inst = repository[instanceId] ?: run {
+        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS instance not found: $instanceId")
+        promise.reject("TTS_ERROR", "TTS instance not found: $instanceId")
+        return
+      }
+      if (!inst.hasEngine()) {
+        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS not initialized")
+        promise.reject("TTS_ERROR", "TTS not initialized")
+        return
+      }
+      promise.resolve(inst.dispatchSampleRate().toDouble())
+    } catch (e: Exception) {
+      Log.e("SherpaOnnxTts", "TTS_ERROR: Failed to get sample rate", e)
+      promise.reject("TTS_ERROR", "Failed to get sample rate", e)
+    }
+  }
+
+  fun getTtsNumSpeakers(instanceId: String, promise: Promise) {
+    try {
+      val inst = repository[instanceId] ?: run {
+        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS instance not found: $instanceId")
+        promise.reject("TTS_ERROR", "TTS instance not found: $instanceId")
+        return
+      }
+      if (!inst.hasEngine()) {
+        Log.e("SherpaOnnxTts", "TTS_ERROR: TTS not initialized")
+        promise.reject("TTS_ERROR", "TTS not initialized")
+        return
+      }
+      promise.resolve(inst.dispatchNumSpeakers().toDouble())
+    } catch (e: Exception) {
+      Log.e("SherpaOnnxTts", "TTS_ERROR: Failed to get number of speakers", e)
+      promise.reject("TTS_ERROR", "Failed to get number of speakers", e)
+    }
+  }
+
+  fun unloadTts(instanceId: String, promise: Promise) {
+    try {
+      val inst = repository.remove(instanceId)
+      if (inst != null) {
+        inst.stopPcmPlayer()
+        inst.releaseEngines()
+      }
+      promise.resolve(null)
+    } catch (e: Exception) {
+      Log.e("SherpaOnnxTts", "TTS_RELEASE_ERROR: Failed to release TTS resources", e)
+      promise.reject("TTS_RELEASE_ERROR", "Failed to release TTS resources", e)
+    }
+  }
+
+  /**
+   * Detect TTS model type and structure without initializing the engine.
+   * Mirrors [SherpaOnnxModule.detectTtsModel] for delegation from facades.
+   */
+  fun detectTtsModel(modelDir: String, modelType: String?, promise: Promise) {
+    try {
+      val result = nativeDetectTtsModel(modelDir, modelType ?: "auto")
+      if (result == null) {
+        Log.e("SherpaOnnx", "DETECT_ERROR: TTS model detection returned null")
+        promise.reject("DETECT_ERROR", "TTS model detection returned null")
+        return
+      }
+      val success = result["success"] as? Boolean ?: false
+      val detectedModels = result["detectedModels"] as? ArrayList<*>
+        ?: arrayListOf<HashMap<String, String>>()
+      val modelTypeStr = result["modelType"] as? String
+      val paths = result["paths"] as? HashMap<*, *>
+
+      val resultMap = Arguments.createMap()
+      resultMap.putBoolean("success", success)
+      val modelsArray = Arguments.createArray()
+      for (model in detectedModels) {
+        val modelMap = model as? HashMap<*, *>
+        if (modelMap != null) {
+          val entry = Arguments.createMap()
+          entry.putString("type", modelMap["type"] as? String ?: "")
+          entry.putString("modelDir", modelMap["modelDir"] as? String ?: "")
+          modelsArray.pushMap(entry)
+        }
+      }
+      resultMap.putArray("detectedModels", modelsArray)
+      if (modelTypeStr != null) {
+        resultMap.putString("modelType", modelTypeStr)
+      }
+      val modelPath = paths?.get("model") as? String
+      if (!modelPath.isNullOrBlank()) {
+        val pathsMap = Arguments.createMap()
+        pathsMap.putString("model", modelPath)
+        resultMap.putMap("paths", pathsMap)
+      }
+      if (!success) {
+        val error = result["error"] as? String
+        if (!error.isNullOrBlank()) {
+          resultMap.putString("error", error)
+        }
+      }
+      val lexiconLanguageCandidates = result["lexiconLanguageCandidates"] as? ArrayList<*>
+      if (!lexiconLanguageCandidates.isNullOrEmpty()) {
+        val candidatesArray = Arguments.createArray()
+        for (c in lexiconLanguageCandidates) {
+          (c as? String)?.let { candidatesArray.pushString(it) }
+        }
+        resultMap.putArray("lexiconLanguageCandidates", candidatesArray)
+      }
+      promise.resolve(resultMap)
+    } catch (e: Exception) {
+      Log.e("SherpaOnnx", "DETECT_ERROR: TTS model detection failed: ${e.message}", e)
+      promise.reject("DETECT_ERROR", "TTS model detection failed: ${e.message}", e)
+    }
+  }
+
+  /**
+   * Batch-resolve TTS catalog metadata from release asset ids (no filesystem).
+   */
+  fun batchTtsCatalogHints(ids: ReadableArray, promise: Promise) {
+    try {
+      val n = ids.size()
+      val arr = Array(n) { i -> ids.getString(i) ?: "" }
+      val raw = nativeBatchTtsCatalogHints(arr)
+      val out = Arguments.createArray()
+      for (m in raw) {
+        val map = Arguments.createMap()
+        map.putString("modelId", m["modelId"] as? String ?: "")
+        map.putString("modelType", m["modelType"] as? String ?: "")
+        val langs = Arguments.createArray()
+        (m["languages"] as? java.util.ArrayList<*>)?.forEach { item ->
+          (item as? String)?.let { langs.pushString(it) }
+        }
+        map.putArray("languages", langs)
+        map.putString("quantization", m["quantization"] as? String ?: "")
+        map.putString("sizeTier", m["sizeTier"] as? String ?: "")
+        out.pushMap(map)
+      }
+      promise.resolve(out)
+    } catch (e: Exception) {
+      Log.e("SherpaOnnx", "batchTtsCatalogHints failed: ${e.message}", e)
+      promise.reject("BATCH_TTS_CATALOG_HINTS_ERROR", e.message, e)
+    }
+  }
+}
