+++ docs/tts-streaming.md
@@ -10,7 +10,7 @@ Incremental speech generation with chunk callbacks: lower time-to-first-byte, pl
 
 - **`ModelPathConfig`** (from `react-native-sherpa-onnx`): `{ type: 'asset' | 'file' | 'auto', path: string }` — directory that contains the TTS model files.
 - **Downloaded models:** use the [Download Manager](download-manager.md) with **`ModelCategory.Tts`**. Valid **`modelId`** values and the GitHub release tag are listed in [Model ids](download-manager.md#model-ids) (`tts-models`).
-- **`detectTtsModel()`** below validates layout and returns detected stacks without loading the engine.
+- **`detectTtsModel()`** below scans the model directory and returns kinds **without** initializing the engine (see [Detection](#detection)).
 
 ## Quick Start
 
@@ -33,11 +33,13 @@ const controller = await tts.generateSpeechStream(
     onChunk: (chunk) => {
       if (chunk.samples.length) void tts.writePcmChunk(chunk.samples);
     },
-    onEnd: () => {
+    onEnd: (e) => {
       void tts.stopPcmPlayer();
+      // e.cancelled is true if the stream ended after cancellation
     },
-    onError: () => {
+    onError: (e) => {
       void tts.stopPcmPlayer();
+      // e.message — native error string
     },
   }
 );
@@ -47,6 +49,30 @@ await tts.destroy();
 ```
 
 **Note:** Zipvoice cloning is **not** supported in streaming on Android; Pocket cloning uses `voiceClone: { kind: 'pocket', referenceAudio: { samples, sampleRate } }`. For Zipvoice voice cloning use batch **`generateSpeech`** on the offline path — [tts-offline.md](tts-offline.md).
+
+### 2) Detect first, then streaming (explicit `modelType` / `modelOptions`)
+
+Same idea as [Offline TTS Quick Start §1](tts-offline.md#1-batch-create-engine-synthesize-destroy): call **`detectTtsModel`** first so you get a narrowed **`modelType`** (and optional Kokoro/Kitten **`lexiconLanguageCandidates`**) without loading the synthesizer, then pass **`modelType`** into **`createStreamingTTS`** when you need engine-specific **`modelOptions`**.
+
+```ts
+import { createStreamingTTS, detectTtsModel } from 'react-native-sherpa-onnx/tts';
+
+const det = await detectTtsModel({ type: 'asset', path: 'models/my-tts-model' });
+if (!det.success || !det.modelType) {
+  throw new Error(det.error ?? 'TTS model detection failed');
+}
+
+const tts = await createStreamingTTS({
+  modelPath: { type: 'asset', path: 'models/my-tts-model' },
+  modelType: det.modelType,
+  numThreads: 2,
+});
+
+// … generateSpeechStream, PCM player, destroy — as in §1 above
+await tts.destroy();
+```
+
+`modelType: 'auto'` (or omitted `modelType`) on **`createStreamingTTS`** still auto-detects on init, but **`modelOptions`** is not available in that mode — set an explicit `modelType` (e.g. from detection) first.
 
 ## Setup (iOS & Android)
 
@@ -72,12 +98,17 @@ function detectTtsModel(
   success: boolean;
   error?: string;
   detectedModels: TtsDetectedModelEntry[];
-  modelType?: TTSModelType | string;
+  modelType?: TTSModelType;
+  // only available for Kokoro/Kitten; otherwise empty
   lexiconLanguageCandidates?: string[];
+  /** When present, how native code chose the model kind (e.g. `fileListing` after a scan, `dirName` from the folder name, `fallbackOrder`, `explicitModelType`, or `nameOnly` for the empty-file-list test path). */
+  detectionSources?: readonly TtsDetectionSource[];
 }>;
 ```
 
-File-based detection and validation without loading the engine; Kokoro/Kitten may return `lexiconLanguageCandidates` for language UI.
+File-based detection and validation **without** initializing the TTS engine: no native synthesizer is created, so this call is comparatively cheap and suitable as a **pre-check** before **`createStreamingTTS`** or **`createTTS`** — for example to obtain a concrete `modelType` (and Kokoro/Kitten `lexiconLanguageCandidates`) so you can pass the right `modelOptions` on init.
+
+**`detectionSources`** is an optional, ordered trace of mechanisms used (stable string literals; see `TtsDetectionSource` in `react-native-sherpa-onnx/tts`). The host-only **name-only** path (empty file list in C++ tests) never validates paths and is not used by production `detectTtsModel` after a real directory scan.
 
 ```ts
 const result = await detectTtsModel({ type: 'asset', path: 'models/my-tts-model' });
@@ -114,9 +145,54 @@ Starts streaming synthesis; events are delivered to `handlers`. Only one stream 
 
 ```ts
 const ctrl = await tts.generateSpeechStream('Hi', undefined, {
-  onChunk: (c) => {},
-  onEnd: () => {},
+  onChunk: (c) => {
+    // c.samples, c.sampleRate, c.progress, c.isFinal
+  },
+  onEnd: (e) => {
+    // e.cancelled
+  },
+  onError: (e) => {
+    // e.message
+  },
 });
+```
+
+### `tts.generateSpeechStreamToFile(text, options, fileOptions, handlers)`
+
+```ts
+generateSpeechStreamToFile(
+  text: string,
+  options: TtsGenerationOptions | undefined,
+  fileOptions: TtsStreamToFileOptions,
+  handlers: TtsStreamToFileHandlers
+): Promise<TtsStreamFileController>;
+```
+
+Streams synthesis directly to a native file sink (v1: WAV). Use this mode for long texts to avoid collecting all samples in JS memory.
+
+```ts
+const ctrl = await tts.generateSpeechStreamToFile(
+  'Save this stream.',
+  undefined,
+  {
+    output: { kind: 'file', path: '/absolute/path/out.wav' },
+    format: 'wav',
+    keepPartialOnCancel: false,
+    emitChunks: true,
+  },
+  {
+    onChunk: (c) => {
+      // optional when emitChunks=true
+      void tts.writePcmChunk(c.samples);
+    },
+    onEnd: (e) => {
+      // e.path, e.bytesWritten, e.sampleRate, e.cancelled
+    },
+    onError: (e) => {
+      // e.message, e.path?
+    },
+  }
+);
 ```
 
 ### `tts.cancelSpeechStream()`
@@ -232,11 +308,17 @@ controller.unsubscribe();
 | `TtsSubtitleItem` | `{ text, start, end }` (seconds) |
 | `TTSModelInfo` | `{ sampleRate, numSpeakers }` |
 | `TtsDetectedModelEntry` | `{ type: TTSModelType \| string; modelDir: string }` |
-| `TtsStreamChunk` | Streaming chunk payload |
-| `TtsStreamEnd` | `{ cancelled: boolean }` + optional ids |
-| `TtsStreamError` | `{ message: string }` + optional ids |
-| `TtsStreamHandlers` | `{ onChunk?, onEnd?, onError? }` |
+| `TtsStreamChunk` | `{ samples, sampleRate, progress, isFinal }` + optional `instanceId`, `requestId` |
+| `TtsStreamEnd` | `{ cancelled: boolean }` + optional `instanceId`, `requestId` |
+| `TtsStreamError` | `{ message: string }` + optional `instanceId`, `requestId` |
+| `TtsStreamHandlers` | `{ onChunk?: (chunk: TtsStreamChunk) => void; onEnd?: (e: TtsStreamEnd) => void; onError?: (e: TtsStreamError) => void }` |
 | `TtsStreamController` | `cancel()`, `unsubscribe()` |
+| `TtsStreamFileOutput` | `{ kind: 'file'; path: string }` |
+| `TtsStreamToFileOptions` | `{ output, format?: 'wav', keepPartialOnCancel?, emitChunks? }` |
+| `TtsStreamFileEnd` | `{ path, bytesWritten, sampleRate, cancelled }` + optional ids |
+| `TtsStreamFileError` | `{ message, path? }` + optional ids |
+| `TtsStreamToFileHandlers` | `{ onChunk?, onEnd?: (e: TtsStreamFileEnd), onError?: (e: TtsStreamFileError) }` |
+| `TtsStreamFileController` | `cancel()`, `unsubscribe()` |
 | `ModelPathConfig` | From `react-native-sherpa-onnx` |
 
 ### Init, update, generation
@@ -275,6 +357,7 @@ Breaking type history: [migration.md](migration.md) → **Text-to-Speech: strict
 | Init throws with `modelOptions` | `modelType` is `'auto'` or omitted | Set explicit `modelType` before passing `modelOptions` |
 | Methods throw after `destroy` | Engine already released | Create a new engine |
 | Streaming error on Android + Zipvoice + ref audio | Unsupported combination | Use batch `generateSpeech` for Zipvoice cloning |
+| `generateSpeechStreamToFile` fails with format error | Non-WAV format requested | Use `format: 'wav'` for v1 stream-to-file |
 | Wrong or slow inference | Provider not built / unavailable | Check [execution-providers.md](execution-providers.md) and native logs |
 
 ## Mapping to Native API
