+++ ios/tts/native/sherpa-onnx-tts-wrapper.mm
@@ -2,7 +2,7 @@
  * sherpa-onnx-tts-wrapper.mm
  *
  * Purpose: Wraps the sherpa-onnx C++ OfflineTts for iOS. Builds config from TtsModelPaths, creates
- * TTS instance, generates audio from text. Used by SherpaOnnx+TTS.mm.
+ * TTS instance, generates audio from text. Used by SherpaOnnx+TTS{Init,Batch,Stream,...}.mm.
  */
 
 #include "sherpa-onnx-tts-wrapper.h"
